remove tests submodule. Update artifact and package name. Update license. Update implementation.
