+++ .travis.yml
@@ -1,5 +1,5 @@
 language: java
 jdk:
   - oraclejdk8
-script: mvn integration-test
+script: mvn install
 install: true
