--- core/pom.xml
@@ -1,160 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    This file is part of Everit - HTTP Basic Authentication.
-
-    Everit - HTTP Basic Authentication is free software: you can redistribute it and/or modify
-    it under the terms of the GNU Lesser General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
-
-    Everit - HTTP Basic Authentication is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU Lesser General Public License for more details.
-
-    You should have received a copy of the GNU Lesser General Public License
-    along with Everit - HTTP Basic Authentication.  If not, see <http://www.gnu.org/licenses/>.
-
--->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>org.everit.config</groupId>
-    <artifactId>org.everit.config.oss</artifactId>
-    <version>5.0.2</version>
-  </parent>
-
-  <groupId>org.everit.osgi</groupId>
-  <artifactId>org.everit.osgi.authentication.http.basic</artifactId>
-  <version>2.0.1</version>
-
-  <packaging>bundle</packaging>
-
-  <name>Everit - HTTP Basic Authentication</name>
-  <description>A HTTP Basic Authentication based on authenticator-api, resource-resolver and authentication-context-api</description>
-
-  <properties>
-    <projectpath>authentication-basic</projectpath>
-    <maven.compiler.source>1.8</maven.compiler.source>
-    <maven.compiler.target>1.8</maven.compiler.target>
-  </properties>
-
-  <scm>
-    <connection>scm:git:git://github.com/everit-org/${projectpath}.git</connection>
-    <developerConnection>scm:git:https://github.com/everit-org/${projectpath}.git</developerConnection>
-    <url>https://github.com/everit-org/${projectpath}</url>
-  </scm>
-
-  <url>http://everit.org</url>
-
-  <issueManagement>
-    <system>GitHub</system>
-    <url>https://github.com/everit-org/${projectpath}/issues</url>
-  </issueManagement>
-
-  <licenses>
-    <license>
-      <name>GNU Lesser General Public License v3.0</name>
-      <url>http://opensource.org/licenses/LGPL-3.0</url>
-      <distribution>repo</distribution>
-    </license>
-  </licenses>
-
-  <developers>
-    <developer>
-      <name>Everit Team</name>
-    </developer>
-  </developers>
-
-  <organization>
-    <name>Everit Kft.</name>
-    <url>http://www.everit.biz</url>
-  </organization>
-
-  <build>
-    <plugins>
-      <plugin>
-        <artifactId>maven-checkstyle-plugin</artifactId>
-        <configuration>
-          <!-- Temporarily added to avoid checkstyle java8 failures -->
-          <failOnViolation>false</failOnViolation>
-        </configuration>
-      </plugin>
-      <plugin>
-        <groupId>org.apache.felix</groupId>
-        <artifactId>maven-bundle-plugin</artifactId>
-        <version>2.5.0</version>
-        <extensions>true</extensions>
-        <configuration>
-          <instructions>
-            <Bundle-SymbolicName>${project.artifactId}</Bundle-SymbolicName>
-            <Import-Package>
-              *
-            </Import-Package>
-            <Export-Package>
-            </Export-Package>
-          </instructions>
-        </configuration>
-      </plugin>
-      <plugin>
-        <groupId>org.apache.felix</groupId>
-        <artifactId>maven-scr-plugin</artifactId>
-        <version>1.17.0</version>
-        <executions>
-          <execution>
-            <id>generate-scr-scrdescriptor</id>
-            <goals>
-              <goal>scr</goal>
-            </goals>
-          </execution>
-        </executions>
-      </plugin>
-    </plugins>
-  </build>
-
-  <dependencies>
-    <dependency>
-      <groupId>org.osgi</groupId>
-      <artifactId>org.osgi.core</artifactId>
-      <version>4.3.1</version>
-      <scope>provided</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.osgi</groupId>
-      <artifactId>org.osgi.compendium</artifactId>
-      <version>4.3.1</version>
-      <scope>provided</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.felix</groupId>
-      <artifactId>org.apache.felix.scr.annotations</artifactId>
-      <version>1.9.8</version>
-    </dependency>
-    <dependency>
-      <groupId>javax.servlet</groupId>
-      <artifactId>javax.servlet-api</artifactId>
-      <version>3.1.0</version>
-      <scope>provided</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.everit.osgi</groupId>
-      <artifactId>org.everit.osgi.authenticator.api</artifactId>
-      <version>1.0.0</version>
-    </dependency>
-    <dependency>
-      <groupId>org.everit.osgi</groupId>
-      <artifactId>org.everit.osgi.resource.resolver.api</artifactId>
-      <version>2.0.0</version>
-    </dependency>
-    <dependency>
-      <groupId>org.everit.osgi</groupId>
-      <artifactId>org.everit.osgi.authentication.context.api</artifactId>
-      <version>1.0.0</version>
-    </dependency>
-  </dependencies>
-
-</project>
