--- core/src/main/resources/OSGI-INF/metatype/metatype.properties
@@ -1,31 +1,0 @@
-#
-# This file is part of Everit - HTTP Basic Authentication.
-#
-# Everit - HTTP Basic Authentication is free software: you can redistribute it and/or modify
-# it under the terms of the GNU Lesser General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# Everit - HTTP Basic Authentication is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU Lesser General Public License for more details.
-#
-# You should have received a copy of the GNU Lesser General Public License
-# along with Everit - HTTP Basic Authentication.  If not, see <http://www.gnu.org/licenses/>.
-#
-
-org.everit.osgi.authentication.http.basic.HttpBasicAuthenticationFilter.name=Everit HTTP Basic Authentication Filter
-org.everit.osgi.authentication.http.basic.HttpBasicAuthenticationFilter.description=The component that implements HTTP Basic Authentication mechanism as a Servlet Filter.
-service.description.name=Service Description
-service.description.description=The description of this component configuration. It is used to easily identify the service registered by this component.
-realm.name=Realm
-realm.description=Pages in the same realm should share credentials. For more information see RFC 1945 (HTTP/1.0) and RFC 2617 (HTTP Authentication referenced by HTTP/1.1).
-authenticator.target.name=Authenticator OSGi filter
-authenticator.target.description=OSGi Service filter expression for Authenticator instance
-resourceIdResolver.target.name=Resource ID Resolver OSGi filter
-resourceIdResolver.target.description=OSGi Service filter expression for ResourceIdResolver instance
-authenticationPropagator.target.name=Authentication Propagator OSGi filter
-authenticationPropagator.target.description=OSGi Service filter expression for AuthenticationPropagator instance
-logService.target.name=Log Service OSGi filter
-logService.target.description=OSGi Service filter expression for LogService instance
