--- tests/src/main/java/org/everit/osgi/authentication/http/basic/tests/sample/HelloWorldServletComponent.java
@@ -1,62 +1,0 @@
-/**
- * This file is part of Everit - HTTP Basic Authentication Tests.
- *
- * Everit - HTTP Basic Authentication Tests is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * Everit - HTTP Basic Authentication Tests is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with Everit - HTTP Basic Authentication Tests.  If not, see <http://www.gnu.org/licenses/>.
- */
-package org.everit.osgi.authentication.http.basic.tests.sample;
-
-import java.io.IOException;
-import java.io.PrintWriter;
-
-import javax.servlet.Servlet;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-import org.apache.felix.scr.annotations.Component;
-import org.apache.felix.scr.annotations.ConfigurationPolicy;
-import org.apache.felix.scr.annotations.Properties;
-import org.apache.felix.scr.annotations.Property;
-import org.apache.felix.scr.annotations.Reference;
-import org.apache.felix.scr.annotations.Service;
-import org.everit.osgi.authentication.context.AuthenticationContext;
-
-@Component(name = "HelloWorldServletComponent", metatype = true,
-        configurationFactory = true, policy = ConfigurationPolicy.REQUIRE)
-@Properties({
-        @Property(name = "authenticationContext.target")
-})
-@Service(value = Servlet.class)
-public class HelloWorldServletComponent extends HttpServlet {
-
-    private static final long serialVersionUID = -5545883781165913751L;
-
-    @Reference(bind = "setAuthenticationContext")
-    private AuthenticationContext authenticationContext;
-
-    @Override
-    protected void doGet(final HttpServletRequest req, final HttpServletResponse resp) throws ServletException,
-            IOException {
-        long currentResourceId = authenticationContext.getCurrentResourceId();
-        resp.setContentType("text/plain");
-        PrintWriter out = resp.getWriter();
-        out.print(currentResourceId);
-    }
-
-    public void setAuthenticationContext(final AuthenticationContext authenticationContext) {
-        this.authenticationContext = authenticationContext;
-    }
-
-}
